requirements.txt
Flask
PyPDF2
python-docx
langdetect
nltk
scikit-learn
pandas 